+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Uliweb Change Log
 * Fix expose for two different view functions but with the same URL does not keep the last one bug
 * Fix rabc init not using functions.has_role and functions.has_permission.
 * Fix rules.py bug
+* Fix rule process, add expose(replace, template) support
 
 0.2.1 Version
 -----------------